done minimum functionality

+++ src/main/java/ru/otus/bbpax/service/QuestionLoader.java
@@ -3,6 +3,6 @@ package ru.otus.bbpax.service;
 import java.util.List;
 
 public interface QuestionLoader<T> {
-    T getQuestionById(long id);
+    T getQuestionById(int id);
     List<T> getQuestions();
 }
